+++ src/components/icons/index.ts
@@ -1,0 +1,3 @@
+export { default as MicrophoneIcon } from './MicrophoneIcon';
+export { default as TranscriptionIcon } from './TranscriptionIcon';
+export { default as CancelIcon } from './CancelIcon';
