+++ src/components/settings/ShowOverlay.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Dropdown } from "../ui/Dropdown";
 import { SettingContainer } from "../ui/SettingContainer";
 import { useSettings } from "../../hooks/useSettings";
+import type { OverlayPosition } from "../../lib/types";
 
 interface ShowOverlayProps {
   descriptionMode?: "inline" | "tooltip";
@@ -20,7 +21,7 @@ export const ShowOverlay: React.FC<ShowOverlayProps> = ({
 }) => {
   const { getSetting, updateSetting, isUpdating } = useSettings();
 
-  const selectedPosition = getSetting("overlay_position") || "bottom";
+  const selectedPosition = (getSetting("overlay_position") || "bottom") as OverlayPosition;
 
   return (
     <SettingContainer
@@ -32,7 +33,7 @@ export const ShowOverlay: React.FC<ShowOverlayProps> = ({
       <Dropdown
         options={overlayOptions}
         selectedValue={selectedPosition}
-        onSelect={(value) => updateSetting("overlay_position", value)}
+        onSelect={(value) => updateSetting("overlay_position", value as OverlayPosition)}
         disabled={isUpdating("overlay_position")}
       />
     </SettingContainer>
