+++ src/lib/types.ts
@@ -19,6 +19,9 @@ export const AudioDeviceSchema = z.object({
   is_default: z.boolean(),
 });
 
+export const OverlayPositionSchema = z.enum(["none", "top", "bottom"]);
+export type OverlayPosition = z.infer<typeof OverlayPositionSchema>;
+
 export const SettingsSchema = z.object({
   bindings: ShortcutBindingsMapSchema,
   push_to_talk: z.boolean(),
@@ -29,7 +32,7 @@ export const SettingsSchema = z.object({
   selected_output_device: z.string().nullable().optional(),
   translate_to_english: z.boolean(),
   selected_language: z.string(),
-  overlay_position: z.string(),
+  overlay_position: OverlayPositionSchema,
   debug_mode: z.boolean(),
 });
